UI: Move widgets cache to widget service.
