+++ ui-ngx/src/app/core/http/widget.service.ts
@@ -16,7 +16,7 @@
 
 import { Injectable } from '@angular/core';
 import { defaultHttpOptionsFromConfig, RequestConfig } from './http-utils';
-import { Observable, of, ReplaySubject, Subject } from 'rxjs';
+import { Observable, of, ReplaySubject } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { PageLink } from '@shared/models/page/page-link';
 import { PageData } from '@shared/models/page/page-data';
@@ -43,14 +43,13 @@ import { ActivationEnd, Router } from '@angular/router';
 })
 export class WidgetService {
 
-  private widgetTypeUpdatedSubject = new Subject<WidgetType>();
-  private widgetsBundleDeletedSubject = new Subject<WidgetsBundle>();
-
   private allWidgetsBundles: Array<WidgetsBundle>;
   private systemWidgetsBundles: Array<WidgetsBundle>;
   private tenantWidgetsBundles: Array<WidgetsBundle>;
 
   private widgetTypeInfosCache = new Map<string, Array<WidgetTypeInfo>>();
+
+  private widgetsInfoInMemoryCache = new Map<string, WidgetInfo>();
 
   private loadWidgetsBundleCacheSubject: ReplaySubject<any>;
 
@@ -117,7 +116,7 @@ export class WidgetService {
           defaultHttpOptionsFromConfig(config)).pipe(
           tap(() => {
             this.invalidateWidgetsBundleCache();
-            this.widgetsBundleDeletedSubject.next(widgetsBundle);
+            this.widgetsBundleDeleted(widgetsBundle);
           })
         );
       }
@@ -217,7 +216,7 @@ export class WidgetService {
     return this.http.post<WidgetTypeDetails>('/api/widgetType', widgetTypeDetails,
       defaultHttpOptionsFromConfig(config)).pipe(
       tap((savedWidgetType) => {
-        this.widgetTypeUpdatedSubject.next(savedWidgetType);
+        this.widgetTypeUpdated(savedWidgetType);
       }));
   }
 
@@ -226,7 +225,7 @@ export class WidgetService {
     return this.http.post<WidgetTypeDetails>('/api/widgetType', widgetTypeDetails,
       defaultHttpOptionsFromConfig(config)).pipe(
       tap((savedWidgetType) => {
-        this.widgetTypeUpdatedSubject.next(savedWidgetType);
+        this.widgetTypeUpdated(savedWidgetType);
       }));
   }
 
@@ -237,7 +236,7 @@ export class WidgetService {
           return this.http.delete(`/api/widgetType/${widgetTypeInstance.id.id}`,
             defaultHttpOptionsFromConfig(config)).pipe(
             tap(() => {
-              this.widgetTypeUpdatedSubject.next(widgetTypeInstance);
+              this.widgetTypeUpdated(widgetTypeInstance);
             })
           );
         }
@@ -263,12 +262,40 @@ export class WidgetService {
       );
   }
 
-  public onWidgetTypeUpdated(): Observable<WidgetType> {
-    return this.widgetTypeUpdatedSubject.asObservable();
+  public createWidgetInfoCacheKey(bundleAlias: string, widgetTypeAlias: string, isSystem: boolean): string {
+    return `${isSystem ? 'sys_' : ''}${bundleAlias}_${widgetTypeAlias}`;
   }
 
-  public onWidgetBundleDeleted(): Observable<WidgetsBundle> {
-    return this.widgetsBundleDeletedSubject.asObservable();
+  public getWidgetInfoFromCache(bundleAlias: string, widgetTypeAlias: string, isSystem: boolean): WidgetInfo | undefined {
+    const key = this.createWidgetInfoCacheKey(bundleAlias, widgetTypeAlias, isSystem);
+    return this.widgetsInfoInMemoryCache.get(key);
+  }
+
+  public putWidgetInfoToCache(widgetInfo: WidgetInfo, bundleAlias: string, widgetTypeAlias: string, isSystem: boolean) {
+    const key = this.createWidgetInfoCacheKey(bundleAlias, widgetTypeAlias, isSystem);
+    this.widgetsInfoInMemoryCache.set(key, widgetInfo);
+  }
+
+  private widgetTypeUpdated(updatedWidgetType: WidgetType): void {
+    this.deleteWidgetInfoFromCache(updatedWidgetType.bundleAlias, updatedWidgetType.alias, updatedWidgetType.tenantId.id === NULL_UUID);
+  }
+
+  private widgetsBundleDeleted(widgetsBundle: WidgetsBundle): void {
+    this.deleteWidgetsBundleFromCache(widgetsBundle.alias, widgetsBundle.tenantId.id === NULL_UUID);
+  }
+
+  private deleteWidgetInfoFromCache(bundleAlias: string, widgetTypeAlias: string, isSystem: boolean) {
+    const key = this.createWidgetInfoCacheKey(bundleAlias, widgetTypeAlias, isSystem);
+    this.widgetsInfoInMemoryCache.delete(key);
+  }
+
+  private deleteWidgetsBundleFromCache(bundleAlias: string, isSystem: boolean) {
+    const key = (isSystem ? 'sys_' : '') + bundleAlias;
+    this.widgetsInfoInMemoryCache.forEach((widgetInfo, cacheKey) => {
+      if (cacheKey.startsWith(key)) {
+        this.widgetsInfoInMemoryCache.delete(cacheKey);
+      }
+    });
   }
 
   private loadWidgetsBundleCache(config?: RequestConfig): Observable<any> {
